+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Usage:
 
 ### 1.19-SNAPSHOT
 * Fix #3591: Fix windows line endings for yaml literal blocks, json serialization config updated
+* Fix #3781: Native Generator does not work in Windows system, always finds multiple native executable
 
 ### 1.18.2 (2025-11-03)
 * Fix #3750: Remove unneeded XMLUtil.createNewDocument method